Add fix to NotePad++ bugs
Since tabs never start in a YAML document line, this fix is quite safe. Some users of PocketMine-MP may use NotePad++ without knowing this bug. However, you may not want to merge this because tabs not at the beginning of a line will also be replaced (and I am too bad at RegExp to fix this)